refactor(phase-2e-residue): full LoC+file-cap lockdown for 9 residue sub-clusters
Drain all 54 ESLint violations across 22 files in 9 sub-clusters
(AccountResolve, BalanceResolve, Browser, Credentials, Home, Otp,
Scrape, Terminate, Timing). Satisfies max-statements:10,
max-lines-per-function:10, max-lines:150, max-len:100, max-params:3,
max-depth:1 simultaneously.

Strategy:
- Function-by-function helper extraction (bundled IXxxArgs interfaces
  for >=4-param call sites, lifted const sentinels for hoisted failure
  shapes, recursive walkers replacing for-of+await loops).
- File-split the four largest files into responsibility siblings +
  slim barrel that preserves every public/test-imported symbol:
    Home/HomeActions.ts -> .Validate .Signal .Modal .Navigate + barrel
    AccountResolve/AccountFromPool.ts -> .Types .Picker .Request + barrel
    AccountResolve/AccountResolveActions.ts -> .Wait .Failures .Classify
        .Discovery .Post .Final + barrel
    BalanceResolve/BalanceResolveActions.ts -> .Shared .Pre .Fetch
        .Dispatch .Extract .Run .Post .Final + barrel
- Tightened ScrapeUiTrigger to derive TrafficHit via
  Awaited<ReturnType<...>> rather than importing IDiscoveredEndpoint,
  satisfying the R-NET-SCRAPE architecture rule.

Verification: ESLint clean across all 9 clusters, tsc --noEmit clean,
test:pipeline 5062/5062 green.
