+++ src/Scrapers/Pipeline/Mediator/AccountResolve/AccountResolveActions.ts
@@ -1,615 +1,21 @@
 /**
- * ACCOUNT-RESOLVE phase Mediator — PRE/ACTION/POST/FINAL.
- * Phase orchestrates ONLY. All logic here.
+ * ACCOUNT-RESOLVE phase Mediator — barrel that composes the five
+ * responsibility siblings (Wait / Failures / Classify / Discovery /
+ * Post / Final) into the single import surface consumed by the
+ * AccountResolve phase orchestrator.
  *
- * Strict contract: by the time this phase exits FINAL, every browser
- * bank's `ctx.accountDiscovery` MUST hold at least one id (or POST
- * fails the run loud with `ACCOUNT_RESOLUTION_FAILED`). The phase is
- * the single source-of-truth for "we know the user's accounts" — the
- * downstream DASHBOARD/SCRAPE phases consume the option without
- * re-running discovery.
+ * Strict contract: by FINAL exit, every browser bank's
+ * `ctx.accountDiscovery` MUST hold at least one id (or POST fails the
+ * run loud with `ACCOUNT_RESOLUTION_FAILED`).
  *
  * PURE GENERIC: works for every bank via the existing 3-source
- * predicate (`discoverAccountsInPool` — response container, GET URL
- * query, POST body). No per-bank branches anywhere in this file.
- *
- * PRE:    blocks on `network.waitForFirstId(20s)` so id-bearing
- *         captures landing late in auth still get into the pool;
- *         emits pool-size telemetry. Full mediator access.
- * ACTION: no-op (sealed action context — no `mediator`). Required
- *         override for the BasePhase template; the real work runs
- *         in PRE/POST where the network surface is reachable.
- * POST:   commits `ctx.accountDiscovery` from the pre-nav pool;
- *         fails loud when ids stay empty.
- * FINAL:  emits resolution telemetry; idempotent.
+ * predicate (`discoverAccountsInPool`).
  */
-
-import { ScraperErrorTypes } from '../../../Base/ErrorTypes.js';
-import { isSome, some } from '../../Types/Option.js';
-import type {
-  IAccountDiscovery,
-  IActionContext,
-  IBillingCycleCatalog,
-  IPipelineContext,
-} from '../../Types/PipelineContext.js';
-import type { Procedure } from '../../Types/Procedure.js';
-import { fail, succeed } from '../../Types/Procedure.js';
-import type { IElementMediator } from '../Elements/ElementMediator.js';
-import type { IDiscoveredEndpoint } from '../Network/NetworkDiscoveryTypes.js';
-import { ACCOUNT_RESOLVE_BUDGET_MS } from '../Timing/TimingConfig.js';
-import { discoverAccountsInPool, poolMaxContainer } from './AccountFromPool.js';
-import { detectBillingCycleCatalog } from './BillingCycleCatalogDetector.js';
-
-/**
- * True when MOCK_MODE is active — lets ACCOUNT-RESOLVE skip its
- * fail-loud checks for the offline snapshot suite, which has no
- * captured network traffic and therefore cannot satisfy the
- * "every bank publishes ids" invariant. Mirrors the
- * `isMockModeOtpActive` valve in OtpFillPhaseActions.
- */
-const isMockModeAccountResolveActive =
-  process.env.MOCK_MODE === '1' || process.env.MOCK_MODE === 'true';
-
-/** Outcome label lookup for the wait result (no ternary per project rules). */
-const WAIT_OUTCOME: Record<'true' | 'false', 'matched' | 'timeout'> = {
-  true: 'matched',
-  false: 'timeout',
-};
-
-/**
- * Caller-owned shape predicate for {@link INetworkDiscovery.waitForFirstId}.
- * Wraps {@link discoverAccountsInPool} into the
- * `(pool) => endpoint | false` shape the network primitive consumes.
- *
- * <p>This indirection is the dependency-inversion seam: ACCOUNT-RESOLVE
- * owns the shape detector; Network owns the polling primitive.
- * Network has zero AccountResolve imports.
- *
- * @param pool - Captured endpoints from the pre-nav pool.
- * @returns First id-bearing endpoint or false.
- */
-function findFirstIdInPool(pool: readonly IDiscoveredEndpoint[]): IDiscoveredEndpoint | false {
-  if (pool.length === 0) return false;
-  const result = discoverAccountsInPool(pool);
-  if (result.endpoint === false) return false;
-  if (result.ids.length === 0) return false;
-  return result.endpoint;
-}
-
-/**
- * Race the id-bearing capture watcher against the page's natural
- * `networkidle` signal — whichever resolves first wins, then a single
- * final pool scan decides the outcome.
- *
- * <p>Background (2026-05-17, PR #234): the previous busy-poll loop
- * burned the full 20 s budget every time the bank's account API
- * response landed with a non-conforming shape (no parsable ids).
- * VisaCal showed this — `/account/info` was captured (URL match) but
- * the response body had no ids, so the predicate kept returning
- * `false` for 20 s × 2 retries. Racing against `waitForNetworkIdle`
- * lets us fail fast once every in-flight request has responded.
- *
- * <p>SOLID / OCP: reuses both existing primitives
- * ({@link INetworkDiscovery.waitForFirstId} for fast-success +
- * {@link IElementMediator.waitForNetworkIdle} for event-driven
- * settle detection). No new wait logic; just composition.
- *
- * @param mediator - Element mediator (network surface owner).
- * @param log - Pipeline logger.
- * @returns True after telemetry is emitted.
- */
-async function awaitAndLog(
-  mediator: IElementMediator,
-  log: IPipelineContext['logger'],
-): Promise<true> {
-  const start = Date.now();
-  const idMatch = mediator.network.waitForFirstId(ACCOUNT_RESOLVE_BUDGET_MS, findFirstIdInPool);
-  await mediator.raceWithNetworkIdle(idMatch, ACCOUNT_RESOLVE_BUDGET_MS);
-  const pool = mediator.network.getPreNavCaptures();
-  logAwaitOutcome(log, pool, start);
-  return true;
-}
-
-/**
- * Compute the matched-key sentinel from the final pool snapshot. The
- * non-nested wrapper keeps the {@link awaitAndLog} body free of the
- * `String(findFirstIdInPool(...) !== false)` nested call that the
- * project's no-restricted-syntax rule rejects.
- *
- * @param pool - Final pre-nav capture pool after the race resolved.
- * @returns Stringified boolean keyed for {@link WAIT_OUTCOME}.
- */
-function evaluateAwaitOutcome(pool: readonly IDiscoveredEndpoint[]): 'true' | 'false' {
-  const matched = findFirstIdInPool(pool);
-  return String(matched !== false) as 'true' | 'false';
-}
-
-/**
- * Build the structured diagnostic and emit it through the pipeline
- * logger. Combining build + emit here keeps the wait orchestrator
- * down to a single call and stops the `log.debug(buildDiag(...))`
- * nested-call lint violation.
- *
- * @param log - Pipeline logger sink.
- * @param pool - Final pre-nav capture pool.
- * @param start - Wall-clock start time in ms.
- * @returns Always true (sentinel for the chained call site).
- */
-function logAwaitOutcome(
-  log: IPipelineContext['logger'],
-  pool: readonly IDiscoveredEndpoint[],
-  start: number,
-): true {
-  const matchedKey = evaluateAwaitOutcome(pool);
-  const diagnostic = buildAwaitDiagnostic(matchedKey, start, pool.length);
-  log.debug(diagnostic);
-  return true;
-}
-
-/** Diagnostic fields for the post-wait `awaitAndLog` debug payload. */
-interface IAwaitDiagnostic {
-  readonly message: string;
-  readonly elapsedMs: string;
-  readonly poolSize: string;
-}
-
-/**
- * Materialise the structured diagnostic logged by {@link awaitAndLog}.
- * Pulled out so the wait orchestrator stays under the per-function LoC
- * budget and the diag shape is easy to update in one place.
- *
- * @param outcomeKey - String form of the boolean outcome (`'true'` /
- *   `'false'`) — keyed off {@link WAIT_OUTCOME}.
- * @param start - Wall-clock start time in ms (from `Date.now()`).
- * @param poolSize - Final capture-pool size after the wait/race.
- * @returns Debug log payload shape consumed by `log.debug`.
- */
-function buildAwaitDiagnostic(
-  outcomeKey: 'true' | 'false',
-  start: number,
-  poolSize: number,
-): IAwaitDiagnostic {
-  return {
-    message: `account-resolve.pre wait → ${WAIT_OUTCOME[outcomeKey]}`,
-    elapsedMs: String(Date.now() - start),
-    poolSize: String(poolSize),
-  };
-}
-
-/**
- * PRE — block on `waitForFirstId` so late-arriving auth-side id
- * captures (Hapoalim's `?accountId=` GETs) make it into the pool
- * before POST extracts. Fails fast when mediator absent.
- * @param input - Pipeline context.
- * @returns Updated context, or no-mediator fail.
- */
-async function executeAccountResolvePre(
-  input: IPipelineContext,
-): Promise<Procedure<IPipelineContext>> {
-  if (!input.mediator.has) return fail(ScraperErrorTypes.Generic, 'ACCOUNT-RESOLVE: no mediator');
-  const mediator = input.mediator.value;
-  const initialPool = mediator.network.getPreNavCaptures();
-  input.logger.debug({ message: `account-resolve.pre pool=${String(initialPool.length)}` });
-  await awaitAndLog(mediator, input.logger);
-  return succeed(input);
-}
-
-/**
- * ACTION — no-op. The sealed action context has no `mediator`
- * surface, so the real work (wait + extract + commit) runs in PRE
- * and POST. Required override for the BasePhase template.
- * @param input - Sealed action context.
- * @returns Pass-through success.
- */
-function executeAccountResolveAction(input: IActionContext): Promise<Procedure<IActionContext>> {
-  const passThrough = succeed(input);
-  return Promise.resolve(passThrough);
-}
-
-/**
- * Builds the `ACCOUNT_RESOLUTION_FAILED` failure for the
- * empty-pool branch. Pulled out so the POST handler stays linear and
- * within the project's per-function line budget.
- * @param poolSize - Pre-nav capture count for the diagnostic message.
- * @returns Failure procedure with the fail-loud message.
- */
-function failAccountResolutionFailed(poolSize: number): Procedure<IPipelineContext> {
-  const msg =
-    'ACCOUNT-RESOLVE POST: ACCOUNT_RESOLUTION_FAILED — ' +
-    `pool=${String(poolSize)} captures, no id-bearing`;
-  return fail(ScraperErrorTypes.Generic, msg);
-}
-
-/** Three-way comparison sentinel returned by {@link compareLocale}. */
-type CompareSign = -1 | 0 | 1;
-
-/**
- * Locale-aware comparator wrapping `String.localeCompare`. Sonar S2871
- * requires sorts on strings to use an explicit locale-aware comparator;
- * Rule #15 forbids primitive number returns from Pipeline functions, so
- * the result is narrowed to a {@link CompareSign} sentinel via
- * `Math.sign` — single expression, no branch coverage cost.
- *
- * @param a - First string.
- * @param b - Second string.
- * @returns -1 when a < b, 0 when equal, 1 when a > b.
- */
-function compareLocale(a: string, b: string): CompareSign {
-  const cmp = a.localeCompare(b);
-  return Math.sign(cmp) as CompareSign;
-}
-
-/**
- * Render a per-container count map as a stable diagnostic string.
- * Used by {@link failAccountResolutionIncomplete} so the F2 error
- * message lists every WK container the picker found AND its size,
- * letting the operator see exactly which container the extractor
- * dropped. Carries only WK constant names + integer counts — no
- * record fields, no ids — so PII can never leak through the message.
- * @param containers - Per-WK-name container split from the picked
- *   endpoint's body.
- * @returns Sorted `name:count` joined with `,`, or `none`.
- */
-function renderContainerCounts(
-  containers: Readonly<Record<string, readonly Record<string, unknown>[]>>,
-): string {
-  const names = Object.keys(containers).sort(compareLocale);
-  if (names.length === 0) return 'none';
-  return names.map((name): string => `${name}:${String(containers[name].length)}`).join(',');
-}
-
-/**
- * Builds the `ACCOUNT_RESOLUTION_INCOMPLETE` failure when the
- * picker returned fewer ids than the SUM of every WK container in
- * the picked endpoint's body — proves the bank served more accounts
- * than the extractor surfaced. Halts the run BEFORE DASHBOARD
- * instead of silently scraping a partial list.
- *
- * <p>Phase 7d tightening: the `expected` count is now the SUM
- * across all WK containers in the picked body (not the legacy
- * single-container max), so VisaCal's `cards: [4]` + `bankAccounts:
- * [3]` payload demands 7 ids, not 4.
- *
- * @param resolved - Resolved id count.
- * @param expected - Sum-of-WK-containers across the pool.
- * @param containers - Per-WK-name split for diagnostic detail.
- * @returns Failure procedure with the fail-loud message.
- */
-function failAccountResolutionIncomplete(
-  resolved: number,
-  expected: number,
-  containers: Readonly<Record<string, readonly Record<string, unknown>[]>>,
-): Procedure<IPipelineContext> {
-  const detail = renderContainerCounts(containers);
-  const msg =
-    'ACCOUNT-RESOLVE POST: ACCOUNT_RESOLUTION_INCOMPLETE — ' +
-    `resolved=${String(resolved)}, expected=${String(expected)}, containers={${detail}}`;
-  return fail(ScraperErrorTypes.Generic, msg);
-}
-
-/**
- * Surface the captureIndex of the picker's chosen endpoint, with
- * `0` as the sentinel when no endpoint was picked (root-array
- * fallback or request-side path). Lookup-style replacement for the
- * inline ternary banned by the project's no-restricted-syntax rule.
- * @param endpoint - Picker output (endpoint or false).
- * @returns Capture index, or 0 sentinel.
- */
-function resolveCaptureIndex(
-  endpoint: ReturnType<typeof discoverAccountsInPool>['endpoint'],
-): number {
-  if (endpoint === false) return 0;
-  return endpoint.captureIndex ?? 0;
-}
-
-/**
- * POST — extracts ids from the pre-nav pool, commits
- * `ctx.accountDiscovery`, or fails loud when the resolution is
- * empty or partial.
- *
- * <p>Two failure modes guard the contract that ACCOUNT-RESOLVE is
- * the single source of truth for account info:
- * <ul>
- *   <li>`ACCOUNT_RESOLUTION_FAILED` — no capture surfaced any
- *   id (empty pool or unrecognized shape).</li>
- *   <li>`ACCOUNT_RESOLUTION_INCOMPLETE` — picker returned
- *   fewer ids than another capture clearly carried, evidence that
- *   the bank served more accounts than the extractor surfaced.</li>
- * </ul>
- *
- * @param input - Pipeline context.
- * @returns Updated context with the discovery option populated, or
- *   one of the two fail-loud procedures.
- */
-/** Discriminated outcome of the POST validation pass. */
-type ResolveClassification =
-  | { readonly kind: 'skip' }
-  | { readonly kind: 'failEmpty'; readonly poolSize: number }
-  | {
-      readonly kind: 'failIncomplete';
-      readonly resolved: number;
-      readonly expected: number;
-      readonly containers: Readonly<Record<string, readonly Record<string, unknown>[]>>;
-    }
-  | {
-      readonly kind: 'commit';
-      readonly pool: readonly IDiscoveredEndpoint[];
-      readonly result: ReturnType<typeof discoverAccountsInPool>;
-    };
-
-/**
- * Build the `failIncomplete` classification carrying the picker
- * shortfall counts. Pulled out so {@link classifyAccountResolveResult}
- * stays a thin 3-branch dispatch.
- *
- * @param result - Resolution outcome from `discoverAccountsInPool`.
- * @param expected - Maximum container size across the capture pool.
- * @returns `failIncomplete` classification ready for the dispatcher.
- */
-function buildFailIncompleteClassification(
-  result: ReturnType<typeof discoverAccountsInPool>,
-  expected: number,
-): ResolveClassification {
-  return {
-    kind: 'failIncomplete',
-    resolved: result.ids.length,
-    expected,
-    containers: result.containers,
-  };
-}
-
-/**
- * Classify the `discoverAccountsInPool` outcome into a tagged
- * union: empty pool → `failEmpty`, expected mismatch →
- * `failIncomplete`, otherwise → `commit`.
- *
- * @param pool - Pre-nav capture pool (drives catalog detection).
- * @param result - Resolution outcome from `discoverAccountsInPool`.
- * @returns Discriminated classification used by the dispatcher.
- */
-function classifyAccountResolveResult(
-  pool: readonly IDiscoveredEndpoint[],
-  result: ReturnType<typeof discoverAccountsInPool>,
-): ResolveClassification {
-  if (result.ids.length === 0) return { kind: 'failEmpty', poolSize: pool.length };
-  const expected = poolMaxContainer(pool);
-  if (result.ids.length < expected) return buildFailIncompleteClassification(result, expected);
-  return { kind: 'commit', pool, result };
-}
-
-/**
- * Build the committed success procedure from a classification's
- * `commit` branch — packages the discovery payload, wraps it in
- * `some`, and clones the pipeline context with `accountDiscovery`
- * populated. Pure: no mediator/network access.
- *
- * @param input - Pipeline context that produced the result.
- * @param pool - Pre-nav capture pool (drives catalog detection).
- * @param result - Resolution outcome from `discoverAccountsInPool`.
- * @returns Success procedure carrying the cloned context.
- */
-function buildAccountResolveSuccess(
-  input: IPipelineContext,
-  pool: readonly IDiscoveredEndpoint[],
-  result: ReturnType<typeof discoverAccountsInPool>,
-): Procedure<IPipelineContext> {
-  const captureIndex = resolveCaptureIndex(result.endpoint);
-  const discoveryPayload = buildDiscoveryPayload(pool, result, captureIndex);
-  const accountDiscovery = some(discoveryPayload);
-  return succeed({ ...input, accountDiscovery });
-}
-
-/** Handler map for {@link ResolveClassification} kinds — OCP-style. */
-type ResolveDispatchMap = {
-  readonly [K in ResolveClassification['kind']]: (
-    input: IPipelineContext,
-    classification: Extract<ResolveClassification, { kind: K }>,
-  ) => Procedure<IPipelineContext>;
-};
-
-const RESOLVE_DISPATCH: ResolveDispatchMap = {
-  /**
-   * Skip branch — passes the context through unchanged (mock-mode /
-   * no-mediator escape).
-   * @param input - Pipeline context.
-   * @returns Pass-through success.
-   */
-  skip: (input): Procedure<IPipelineContext> => succeed(input),
-  /**
-   * Empty-pool branch — emits the fail-loud `ACCOUNT_RESOLUTION_FAILED`
-   * with the diagnostic pool size.
-   * @param _input - Unused pipeline context.
-   * @param c - Classification with `poolSize`.
-   * @returns Failure procedure.
-   */
-  failEmpty: (_input, c): Procedure<IPipelineContext> => failAccountResolutionFailed(c.poolSize),
-  /**
-   * Partial-resolution branch — emits the fail-loud
-   * `ACCOUNT_RESOLUTION_INCOMPLETE` with resolved/expected/container
-   * diagnostics.
-   * @param _input - Unused pipeline context.
-   * @param c - Classification with resolved/expected/containers.
-   * @returns Failure procedure.
-   */
-  failIncomplete: (_input, c): Procedure<IPipelineContext> =>
-    failAccountResolutionIncomplete(c.resolved, c.expected, c.containers),
-  /**
-   * Commit branch — packages the discovery payload onto the context.
-   * @param input - Pipeline context.
-   * @param c - Classification carrying pool + result.
-   * @returns Success procedure with `accountDiscovery` populated.
-   */
-  commit: (input, c): Procedure<IPipelineContext> =>
-    buildAccountResolveSuccess(input, c.pool, c.result),
-};
-
-/**
- * Dispatch the classification's `kind` to its concrete procedure via
- * the {@link RESOLVE_DISPATCH} handler map. Single-line body; the
- * `as never` cast lets us index the map with a narrowed kind without
- * losing type safety.
- *
- * @param input - Pipeline context.
- * @param classification - Outcome of {@link classifyAccountResolveResult}.
- * @returns The success / fail procedure for this branch.
- */
-function dispatchAccountResolveClassification(
-  input: IPipelineContext,
-  classification: ResolveClassification,
-): Procedure<IPipelineContext> {
-  const handler = RESOLVE_DISPATCH[classification.kind] as (
-    input: IPipelineContext,
-    classification: ResolveClassification,
-  ) => Procedure<IPipelineContext>;
-  return handler(input, classification);
-}
-
-/**
- * POST — extracts ids from the pre-nav pool, commits
- * `ctx.accountDiscovery`, or fails loud when the resolution is empty
- * or partial via {@link classifyAccountResolveResult} +
- * {@link dispatchAccountResolveClassification}.
- *
- * @param input - Pipeline context.
- * @returns Updated context with the discovery option populated, or
- *   one of the two fail-loud procedures.
- */
-/**
- * Resolve the success/fail procedure from the live mediator state.
- * Pure pull-from-pool + classify + dispatch — pulled out so the
- * outer phase stays a thin gate.
- *
- * @param input - Pipeline context.
- * @param mediator - Unwrapped element mediator (POST guard already
- *   verified `input.mediator.has`).
- * @returns Procedure for this run — success/fail.
- */
-function buildAccountResolvePostResult(
-  input: IPipelineContext,
-  mediator: IElementMediator,
-): Procedure<IPipelineContext> {
-  const pool = mediator.network.getPreNavCaptures();
-  const result = discoverAccountsInPool(pool);
-  const classification = classifyAccountResolveResult(pool, result);
-  return dispatchAccountResolveClassification(input, classification);
-}
-
-/**
- * POST — extracts ids from the pre-nav pool, commits
- * `ctx.accountDiscovery`, or fails loud when the resolution is empty
- * or partial via {@link classifyAccountResolveResult} +
- * {@link dispatchAccountResolveClassification}.
- *
- * @param input - Pipeline context.
- * @returns Updated context with the discovery option populated, or
- *   one of the two fail-loud procedures.
- */
-function executeAccountResolvePost(input: IPipelineContext): Promise<Procedure<IPipelineContext>> {
-  if (!input.mediator.has || isMockModeAccountResolveActive) {
-    const passThrough = succeed(input);
-    return Promise.resolve(passThrough);
-  }
-  const dispatched = buildAccountResolvePostResult(input, input.mediator.value);
-  return Promise.resolve(dispatched);
-}
-
-/** Args bundle for {@link buildDiscoveryPayload}. */
-interface IDiscoveryResult {
-  readonly ids: readonly string[];
-  readonly records: readonly Record<string, unknown>[];
-  readonly containers: Readonly<Record<string, readonly Record<string, unknown>[]>>;
-}
-
-/**
- * Build the core {@link IAccountDiscovery} record (no optional fields).
- * Extracted so {@link buildDiscoveryPayload} stays within the project's
- * 10-line method budget.
- *
- * @param result - Account-resolution outcome from `discoverAccountsInPool`.
- * @param captureIndex - Index of the capture that surfaced the account ids.
- * @returns Base discovery record sans optional catalog.
- */
-function buildBaseDiscovery(result: IDiscoveryResult, captureIndex: number): IAccountDiscovery {
-  return {
-    ids: result.ids,
-    records: result.records,
-    containers: result.containers,
-    endpointCaptureIndex: captureIndex,
-  };
-}
-
-/**
- * Assemble the {@link IAccountDiscovery} payload committed onto
- * `ctx.accountDiscovery`. Adds the optional billing-cycle catalog
- * when the detector recognises a known shape in the pre-nav pool;
- * omits the field for non-cycling banks (current accounts).
- *
- * @param pool - Pre-nav capture pool from the mediator.
- * @param result - Account-resolution outcome from `discoverAccountsInPool`.
- * @param captureIndex - Index of the capture that surfaced the account ids.
- * @returns Fully-populated discovery record ready to wrap in `some()`.
- */
-function buildDiscoveryPayload(
-  pool: readonly IDiscoveredEndpoint[],
-  result: IDiscoveryResult,
-  captureIndex: number,
-): IAccountDiscovery {
-  const base = buildBaseDiscovery(result, captureIndex);
-  const catalogOption = detectBillingCycleCatalog(pool);
-  if (!isSome(catalogOption)) return base;
-  const billingCycleCatalog: IBillingCycleCatalog = catalogOption.value;
-  return { ...base, billingCycleCatalog };
-}
-
-/** First-id label lookup — no `''` fallbacks per project rules. */
-const FIRST_ID_LABEL: Record<'true' | 'false', (ids: readonly string[]) => string> = {
-  /**
-   * Non-empty case — return the head id.
-   * @param ids - Resolved id list.
-   * @returns First id.
-   */
-  true: (ids): string => ids[0],
-  /**
-   * Empty case — sentinel string for telemetry parity.
-   * @returns Sentinel.
-   */
-  false: (): string => 'none',
-};
-
-/**
- * Surface the resolved id list from the discovery option, returning
- * an empty array when the option is `none`. Lookup-style replacement
- * for the inline ternary.
- * @param ctx - Pipeline context.
- * @returns Resolved id list (possibly empty).
- */
-function resolvedIds(ctx: IPipelineContext): readonly string[] {
-  const has = ctx.accountDiscovery.has;
-  if (!has) return [];
-  return ctx.accountDiscovery.value.ids;
-}
-
-/**
- * FINAL — resolution telemetry. Idempotent: re-running this stage
- * with the same context produces the same log line.
- * @param input - Pipeline context.
- * @returns Pass-through success.
- */
-function executeAccountResolveFinal(input: IPipelineContext): Promise<Procedure<IPipelineContext>> {
-  const ids = resolvedIds(input);
-  const labelKey = String(ids.length > 0) as 'true' | 'false';
-  const firstId = FIRST_ID_LABEL[labelKey](ids);
-  input.logger.debug({
-    message: `account-resolve.final ids=${String(ids.length)} firstId=${firstId}`,
-  });
-  const success = succeed(input);
-  return Promise.resolve(success);
-}
 
 export { ACCOUNT_RESOLVE_BUDGET_MS } from '../Timing/TimingConfig.js';
+export { executeAccountResolveFinal } from './AccountResolveActions.Final.js';
+export { executeAccountResolvePost } from './AccountResolveActions.Post.js';
 export {
   executeAccountResolveAction,
-  executeAccountResolveFinal,
-  executeAccountResolvePost,
   executeAccountResolvePre,
-};
+} from './AccountResolveActions.Wait.js';
