+++ src/Scrapers/Pipeline/Mediator/Scrape/BfsFieldSearch/BfsFieldSearch.ts
@@ -1,172 +1,33 @@
 /**
- * BFS field search — breadth-first walk over nested JSON shapes
- * to locate well-known field names.
+ * BFS field search barrel — composes the {@link Match} (scalar
+ * matchers) and {@link Flatten} (tree flattener) siblings into the
+ * `findFieldValue` / `findAllFieldValues` entry points consumed by
+ * TxnMapper, ContainerClaim and AccountExtractor.
  *
  * Extracted from ScrapeAutoMapper as part of the Phase 5
  * pipeline-decoupling split (master plan
  * pipeline-decoupling-master-2026-05-28 / phase-5).
  */
 
-import { ScraperErrorTypes } from '../../../../Base/ErrorTypes.js';
-import type { IFieldMatch } from '../../../Types/FieldMatch.js';
-import type { Procedure } from '../../../Types/Procedure.js';
-import { fail, isOk, succeed } from '../../../Types/Procedure.js';
-import {
-  type ApiRecord,
-  MAX_SEARCH_DEPTH,
-  type ScalarFieldHit,
-  type UntypedValue,
-} from '../AutoMapperFacade/AutoMapperTypes.js';
-
-/** BFS queue item for iterative deep search. */
-interface ISearchItem {
-  readonly value: Record<string, unknown>;
-  readonly depth: number;
-}
+import type { ScalarFieldHit } from '../AutoMapperFacade/AutoMapperTypes.js';
+import { flattenObjectTree, type ISearchItem } from './BfsFieldSearch.Flatten.js';
+import { isSearchableObject, matchField, matchFieldInRecord } from './BfsFieldSearch.Match.js';
 
 /**
- * Check if a value is a searchable object.
- * @param val - Value to check.
- * @returns True if val is a non-null, non-array object.
+ * Find first scalar hit across the flattened object tree.
+ * Pulled out so {@link findFieldValue} stays under the LoC budget.
+ * @param obj - Root object (already missed at the root level).
+ * @param fieldNames - WellKnown field names to try at each level.
+ * @returns First scalar hit or false.
  */
-function isSearchableObject(val: UntypedValue): boolean {
-  return typeof val === 'object' && val !== null && !Array.isArray(val);
-}
-
-/**
- * Try matching one WK name against record keys.
- * @param record - Object to search.
- * @param recordKeys - Pre-computed keys.
- * @param wkName - WellKnown field name.
- * @returns IFieldMatch or false.
- */
-function tryMatchWk(
-  record: Record<string, unknown>,
-  recordKeys: readonly string[],
-  wkName: string,
-): Procedure<IFieldMatch> {
-  const wkLower = wkName.toLowerCase();
-  const originalKey = recordKeys.find((k): boolean => k.toLowerCase() === wkLower);
-  if (!originalKey) return fail(ScraperErrorTypes.Generic, 'key not found');
-  const val = record[originalKey];
-  const isScalar = typeof val === 'string' || typeof val === 'number';
-  if (!isScalar) return fail(ScraperErrorTypes.Generic, 'not scalar');
-  return succeed({ originalKey, value: val, matchingKey: wkName });
-}
-
-/**
- * Case-insensitive field match via Result Pattern.
- * @param record - Object to search.
- * @param fieldNames - WellKnown field names to match against.
- * @returns Procedure with IFieldMatch on success, failure on miss.
- */
-function matchField(
-  record: Record<string, unknown>,
-  fieldNames: readonly string[],
-): Procedure<IFieldMatch> {
-  const recordKeys = Object.keys(record);
-  /**
-   * Try one WK name against this record.
-   * @param wk - WK name.
-   * @returns Match result.
-   */
-  const tryWk = (wk: string): Procedure<IFieldMatch> => tryMatchWk(record, recordKeys, wk);
-  const results = fieldNames.map(tryWk);
-  const hit = results.find(isOk);
-  return hit ?? fail(ScraperErrorTypes.Generic, 'no WK match');
-}
-
-/**
- * Backward-compat wrapper — returns raw value or false.
- * @param record - Object to check.
- * @param fieldNames - WellKnown names.
- * @returns Matched value or false.
- */
-function matchFieldInRecord(
-  record: Record<string, unknown>,
+function findInFlattenedTree(
+  obj: Record<string, unknown>,
   fieldNames: readonly string[],
 ): ScalarFieldHit {
-  const result = matchField(record, fieldNames);
-  if (!isOk(result)) return false;
-  return result.value.value;
-}
-
-/**
- * Build BFS child entries from a record's values, pre-filtered to
- * searchable child objects. Pulled out so {@link enqueueChildren}
- * stays a thin guard + push.
- *
- * @param record - Parent object.
- * @param depth - Current BFS depth.
- * @returns Child entries with depth+1.
- */
-function buildBfsChildren(record: Record<string, unknown>, depth: number): readonly ISearchItem[] {
-  const nextDepth = depth + 1;
-  const recordValues = Object.values(record) as readonly UntypedValue[];
-  return recordValues.filter(isSearchableObject).map(
-    (child): ISearchItem => ({
-      value: child as Record<string, unknown>,
-      depth: nextDepth,
-    }),
-  );
-}
-
-/**
- * Enqueue child objects from a record into BFS queue.
- * @param record - Parent object.
- * @param depth - Current depth.
- * @param queue - BFS queue to append to.
- * @returns True if children were enqueued.
- */
-function enqueueChildren(
-  record: Record<string, unknown>,
-  depth: number,
-  queue: ISearchItem[],
-): boolean {
-  if (depth >= MAX_SEARCH_DEPTH) return false;
-  const children = buildBfsChildren(record, depth);
-  queue.push(...children);
-  return true;
-}
-
-/**
- * Process one BFS level and return next level items.
- * @param items - Current level items.
- * @returns Next level items.
- */
-function bfsOneLevel(items: readonly ISearchItem[]): readonly ISearchItem[] {
-  const next: ISearchItem[] = [];
-  for (const item of items) {
-    enqueueChildren(item.value, item.depth, next);
-  }
-  return next;
-}
-
-/**
- * Recursive BFS level processor — accumulates all objects.
- * @param current - Current level items.
- * @param accum - All items collected so far.
- * @returns Complete flat list.
- */
-function bfsAccumulate(
-  current: readonly ISearchItem[],
-  accum: readonly ISearchItem[],
-): readonly ISearchItem[] {
-  if (current.length === 0) return accum;
-  const merged = [...accum, ...current];
-  const next = bfsOneLevel(current);
-  return bfsAccumulate(next, merged);
-}
-
-/**
- * Flatten a nested object tree into an array of records via BFS.
- * @param root - Root API record.
- * @returns All records found at any depth.
- */
-function flattenObjectTree(root: ApiRecord): readonly ApiRecord[] {
-  const seed: ISearchItem[] = [{ value: root, depth: 0 }];
-  const all = bfsAccumulate(seed, []);
-  return all.map((entry): ApiRecord => entry.value);
+  const allObjects = flattenObjectTree(obj);
+  const results = allObjects.map((o): ScalarFieldHit => matchFieldInRecord(o, fieldNames));
+  const hit = results.find((r): boolean => r !== false);
+  return hit ?? false;
 }
 
 /**
@@ -181,43 +42,41 @@ function findFieldValue(
 ): ScalarFieldHit {
   const rootHit = matchFieldInRecord(obj, fieldNames);
   if (rootHit !== false) return rootHit;
-  const allObjects = flattenObjectTree(obj);
-  const results = allObjects.map((o): ScalarFieldHit => matchFieldInRecord(o, fieldNames));
-  const hit = results.find((r): boolean => r !== false);
-  return hit ?? false;
+  return findInFlattenedTree(obj, fieldNames);
 }
 
 /**
- * Recursive collector that visits every plain-object node reachable
- * through nested objects AND arrays, accumulating
- * {@link matchFieldInRecord} hits as it goes. Counterpart of
- * {@link flattenObjectTree}'s arrays-not-descended limitation: id
- * discovery shapes nest objects inside arrays (e.g. VisaCal
- * `result.bigNumbers[].cards[]`), so the array-skipping flattener
- * silently shadows the children.
- *
- * @param value - Current JSON value (any nesting depth).
+ * Walk an array node by recursing into each item.
+ * Pulled out so {@link collectFieldValuesDeep} stays at depth 1.
+ * @param arr - Array of unknown items.
  * @param fieldNames - WellKnown field names to try at each record.
- * @returns Array of hits collected at this node and below.
+ * @returns Hits collected across all items.
  */
-function collectFieldValuesDeep(
-  value: unknown,
+function collectFromArray(
+  arr: readonly unknown[],
   fieldNames: readonly string[],
 ): readonly ScalarFieldHit[] {
-  if (Array.isArray(value)) {
-    const arr = value as readonly unknown[];
-    return arr.flatMap((item): readonly ScalarFieldHit[] =>
-      collectFieldValuesDeep(item, fieldNames),
-    );
-  }
-  if (value === null || typeof value !== 'object') return [];
-  return collectFieldValuesFromRecord(value as Record<string, unknown>, fieldNames);
+  return arr.flatMap((item): readonly ScalarFieldHit[] => collectFieldValuesDeep(item, fieldNames));
+}
+
+/**
+ * Collect hits from every child value of a record.
+ * Pulled out so {@link collectFieldValuesFromRecord} stays under the LoC budget.
+ * @param record - Record whose children should be descended.
+ * @param fieldNames - WellKnown field names to try at each level.
+ * @returns Hits collected across all descendants.
+ */
+function collectChildHits(
+  record: Record<string, unknown>,
+  fieldNames: readonly string[],
+): readonly ScalarFieldHit[] {
+  const children = Object.values(record);
+  return children.flatMap((c): readonly ScalarFieldHit[] => collectFieldValuesDeep(c, fieldNames));
 }
 
 /**
  * Match a record at this level + recurse into every child value.
  * Hoisted so {@link collectFieldValuesDeep} stays at depth 1.
- *
  * @param record - Record at the current node.
  * @param fieldNames - WellKnown field names to try at each record.
  * @returns Hits collected at this record + descendants.
@@ -228,26 +87,29 @@ function collectFieldValuesFromRecord(
 ): readonly ScalarFieldHit[] {
   const direct = matchFieldInRecord(record, fieldNames);
   const hereHits: readonly ScalarFieldHit[] = direct === false ? [] : [direct];
-  const children = Object.values(record);
-  const childHits = children.flatMap((child): readonly ScalarFieldHit[] =>
-    collectFieldValuesDeep(child, fieldNames),
-  );
+  const childHits = collectChildHits(record, fieldNames);
   return [...hereHits, ...childHits];
+}
+
+/**
+ * Recursive collector that visits every plain-object node reachable
+ * through nested objects AND arrays.
+ * @param value - Current JSON value (any nesting depth).
+ * @param fieldNames - WellKnown field names to try at each record.
+ * @returns Array of hits collected at this node and below.
+ */
+function collectFieldValuesDeep(
+  value: unknown,
+  fieldNames: readonly string[],
+): readonly ScalarFieldHit[] {
+  if (Array.isArray(value)) return collectFromArray(value as readonly unknown[], fieldNames);
+  if (value === null || typeof value !== 'object') return [];
+  return collectFieldValuesFromRecord(value as Record<string, unknown>, fieldNames);
 }
 
 /**
  * Collect every matching field value at any nesting depth — walks
  * both objects and arrays.
- *
- * <p>Counterpart to {@link findFieldValue} that returns ALL hits, not
- * just the first. Required by SCRAPE.final attribution when one
- * captured endpoint exposes multiple WK ids at different depths — for
- * example, VisaCal's `getBigNumberAndDetails` POST body carries a
- * parent `bankAccountUniqueId` on the request side AND a child
- * `cardUniqueId` inside `result.bigNumbers[].cards[]` on the response
- * side. The first-hit variant silently shadows one with the other;
- * the all-hits variant lets the caller match either.
- *
  * @param obj - Root object to search.
  * @param fieldNames - WellKnown field names to try at each level.
  * @returns All scalar hits collected via deep walk (may be empty).
