+++ src/Scrapers/Pipeline/Mediator/Scrape/ScrapeUiTrigger.ts
@@ -14,6 +14,12 @@ import type { Procedure } from '../../Types/Procedure.js';
 import { succeed } from '../../Types/Procedure.js';
 import type { IElementMediator } from '../Elements/ElementMediator.js';
 
+/**
+ * Slim traffic-hit shape derived from the mediator surface without
+ * importing the forbidden `IDiscoveredEndpoint` symbol (R-NET-SCRAPE).
+ */
+type TrafficHit = Awaited<ReturnType<IElementMediator['network']['waitForTraffic']>>;
+
 /** Best-effort timeout — don't block, traffic captured in LOGIN.POST. */
 const TRAFFIC_TIMEOUT = 5000;
 /** Timeout for WK element discovery. */
@@ -41,6 +47,19 @@ async function tryWkClick(
 }
 
 /**
+ * Mask + emit the traffic hit when one arrived. Pulled out so {@link waitAndTrace}
+ * stays under the per-function LoC budget.
+ * @param hit - Traffic capture or false.
+ * @param logger - Pipeline logger.
+ * @returns Procedure carrying whether a hit was logged.
+ */
+function logTrafficHit(hit: TrafficHit, logger?: ScraperLogger): Procedure<boolean> {
+  if (!hit) return succeed(false);
+  logger?.trace({ method: hit.method, url: maskVisibleText(hit.url) });
+  return succeed(true);
+}
+
+/**
  * Wait for traffic after a click, log result.
  * @param mediator - Element mediator.
  * @param label - Clicked element label.
@@ -52,11 +71,10 @@ async function waitAndTrace(
   label: string,
   logger?: ScraperLogger,
 ): Promise<boolean> {
-  const masked = maskVisibleText(`Clicked '${label}'`);
-  logger?.debug({ message: masked });
+  logger?.debug({ message: maskVisibleText(`Clicked '${label}'`) });
   const hit = await mediator.network.waitForTraffic(TXN_PATTERNS, TRAFFIC_TIMEOUT);
-  if (hit) logger?.trace({ method: hit.method, url: maskVisibleText(hit.url) });
-  return Boolean(hit);
+  const logged = logTrafficHit(hit, logger);
+  return logged.success && logged.value;
 }
 
 /**
@@ -80,6 +98,45 @@ async function tryWkGroup(
   return succeed(hasTraffic);
 }
 
+/** Tier names used by {@link tryWkTier} — keeps the orchestrator declarative. */
+const TRIGGER_TIERS = ['TRANSACTIONS', 'MENU_EXPAND'] as const;
+
+/**
+ * Try one WK trigger tier; returns success procedure or false.
+ * @param mediator - Element mediator (black box).
+ * @param tier - Tier key in {@link WK_DASHBOARD}.
+ * @param logger - Pipeline logger.
+ * @returns Procedure success when the click landed, else false.
+ */
+function tryWkTier(
+  mediator: IElementMediator,
+  tier: (typeof TRIGGER_TIERS)[number],
+  logger?: ScraperLogger,
+): Promise<Procedure<boolean> | false> {
+  const group = WK_DASHBOARD[tier] as unknown as readonly SelectorCandidate[];
+  return tryWkGroup(mediator, group, logger);
+}
+
+/** Bundled args for the recursive tier walker — keeps params ≤ 3. */
+interface ITierWalkArgs {
+  readonly mediator: IElementMediator;
+  readonly logger?: ScraperLogger;
+  readonly index: number;
+}
+
+/**
+ * Walk the trigger tiers sequentially without `await` inside a loop.
+ * @param args - Bundled mediator + logger + current tier index.
+ * @returns First successful hit, or false when all tiers exhausted.
+ */
+async function walkTriggerTiers(args: ITierWalkArgs): Promise<Procedure<boolean> | false> {
+  if (args.index >= TRIGGER_TIERS.length) return false;
+  const tier = TRIGGER_TIERS[args.index];
+  const hit = await tryWkTier(args.mediator, tier, args.logger);
+  if (hit !== false) return hit;
+  return walkTriggerTiers({ ...args, index: args.index + 1 });
+}
+
 /**
  * Best-effort TRIGGER: ONE click attempt, short wait, then succeed.
  * @param mediator - Element mediator (black box).
@@ -90,12 +147,8 @@ async function triggerDashboardUi(
   mediator: IElementMediator,
   logger?: ScraperLogger,
 ): Promise<Procedure<boolean>> {
-  const txn = WK_DASHBOARD.TRANSACTIONS as unknown as readonly SelectorCandidate[];
-  const txnHit = await tryWkGroup(mediator, txn, logger);
-  if (txnHit !== false) return txnHit;
-  const menu = WK_DASHBOARD.MENU_EXPAND as unknown as readonly SelectorCandidate[];
-  const menuHit = await tryWkGroup(mediator, menu, logger);
-  if (menuHit !== false) return menuHit;
+  const hit = await walkTriggerTiers({ mediator, logger, index: 0 });
+  if (hit !== false) return hit;
   logger?.debug({ message: 'No UI trigger — traffic from LOGIN' });
   return succeed(false);
 }
